refactoring common batch listener files to common location, as this is being used by multiple components now -- indexing and for ontology graph traversal source file loading
